Update core wasm32 typedefs to be compatible with reference-sysroot.
Change `time_t` from `long` (32-bit) to `long long` (64-bit). This fixes the
[2038 bug](https://en.wikipedia.org/wiki/Year_2038_problem). It's also
consistent with the [x32 ABI](https://en.wikipedia.org/wiki/X32_ABI).

Change the fallback for `wchar_t` and `wint_t` from `unsigned` to `int`. This
is not an ABI change in practice because the headers prefer to use
`__WCHAR_TYPE__` and `__WINT_TYPE__` when they're defined, and clang defines
those to be `int`. So this just makes the fallback cases for other
hypothetical compilers consistent with clang. (And fwiw, no one specifically
chose `int` for these, that's just the default in clang if the target doesn't
override it. So if anyone feels strongly that wchar_t/wint_t should be
something different, we can change them).

Change `suseconds_t` from `long` to `long long`. This matches x32, uses what
would otherwise be padding a `struct timeval`, and generally makes it easier
for users to avoid overflow in time calculations.

Change `_Reg` from `__PTRDIFF_TYPE__` (32-bit) to `__INT64_TYPE__` (64-bit).
`_Reg` is only used for `nlink_t` and `register_t`. For `nlink_t`, this
matches x32, and theoretically supports filesystems that support billions
of links. For `register_t`, since wasm supports native `i64` values, a 64-bit
int is the closest mapping of this concept.

+++ system/lib/libc/musl/arch/emscripten/bits/alltypes.h
@@ -72,12 +72,12 @@ typedef long double double_t;
 #endif
 
 #if defined(__NEED_time_t) && !defined(__DEFINED_time_t)
-typedef long time_t;
+typedef long long time_t;
 #define __DEFINED_time_t
 #endif
 
 #if defined(__NEED_suseconds_t) && !defined(__DEFINED_suseconds_t)
-typedef long suseconds_t;
+typedef long long suseconds_t;
 #define __DEFINED_suseconds_t
 #endif
